[feature/STATS-005] state로 상태관리 및 오류수정 -그래프내 제스쳐 비활성화 - 밀리초 단위 통일

+++ app/src/main/java/com/mukmuk/todori/ui/screen/stats/component/MonthCard.kt
@@ -43,7 +43,7 @@ fun MonthCard(
     val avgHours = avgStudyMinutes / 60
     val avgMinutes = avgStudyMinutes % 60
 
-    val totalStudyMinutes = (totalStudyTimeMillis / 1000 / 60)
+    val totalStudyMinutes = (totalStudyTimeMillis / 1000 / 60).toInt()
     val totalHours = totalStudyMinutes / 60
     val totalMinutes = totalStudyMinutes % 60
